Fix automatic inclusion of all functions.
This is better done in the transpiler and results in less code that is
also more correct.  Add tests for this and refactor the code and tests.
